feat(ui): ask what to edit when a game has no controller layout of its own
Opening the controller editor for a game that only inherits its
console's layout silently started a blank layout, losing the inherited
customizations. Ask instead, matching what launch actually does:
modify the parent layout in place (every game using it changes), seed
a copy bound to this game, or start empty. The editor gains a
seed_from option that loads another layout's content but saves as a
never-written new file under the game's name.

+++ crates/ira/src/ui/play_history.rs
@@ -562,7 +562,7 @@ mod tests {
         // Many games in one day: each gets its own segment and swatch —
         // nothing is lumped into a grey "Other" bucket.
         let s: Vec<ira_models::PlaySession> = (1..=8i64)
-            .map(|i| session(i, i, 600 * (9 - i as i64)))
+            .map(|i| session(i, i, 600 * (9 - i)))
             .collect();
         let mut day: HashMap<i64, Vec<&ira_models::PlaySession>> = HashMap::new();
         for session in &s {
